// front-end/packages/shared/src/models/couponType.ts
/**
 * Generated by orval v7.14.0 🍺
 * Do not edit manually.
 * OpenAPI specification - RING-Bookstore
 * OpenAPI documentation.
 * OpenAPI spec version: 2.0
 */

export type CouponType = (typeof CouponType)[keyof typeof CouponType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CouponType = {
  SHIPPING: "SHIPPING",
  PRODUCT: "PRODUCT",
} as const;